Implemented new DB Helper and changed db schema
Implemented a database helper based on the AndroidSQLiteAssetHelper
helper class. This allows a premade database to be used to construct the
application's database.

Added Location and VenueType classes, to represent the location and
venue_type tables in the database

Modified Venue and Rating to reflect changes to the database schema